+++ modules/rest_blockfrost/src/types.rs
@@ -55,7 +55,7 @@ pub struct DRepUpdateREST {
 #[derive(Serialize)]
 pub struct DRepVoteREST {
     pub tx_hash: String,
-    pub cert_index: u64,
+    pub cert_index: usize,
     pub vote: Vote,
 }
